Remove AssessmentPrintSection, add deskStudy for print
