+++ client/pages/AssessmentSection/AssessmentSection.tsx
@@ -1,13 +1,15 @@
 import './AssessmentSection.scss'
-import React from 'react'
+import React, { useEffect, useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useParams } from 'react-router-dom'
 
 import { AssessmentName } from '@meta/assessment'
 
-import { useAssessmentSection } from '@client/store/assessment'
-import { useTableSections } from '@client/store/pages/assessmentSection'
+import { useAppDispatch } from '@client/store'
+import { useAssessment, useAssessmentSection, useCycle } from '@client/store/assessment'
+import { AssessmentSectionActions, useTableSections } from '@client/store/pages/assessmentSection'
 import { useCanEditSection } from '@client/store/user'
+import { useCountryIso } from '@client/hooks'
 import { useIsPrint } from '@client/hooks/useIsPath'
 
 import DataTable from './DataTable'
@@ -15,16 +17,41 @@ import Descriptions, { GeneralComments } from './Descriptions'
 import SectionHeader from './SectionHeader'
 import Title from './Title'
 
-const AssessmentSection: React.FC = () => {
+type Props = {
+  section?: string
+}
+
+const AssessmentSection: React.FC<Props> = (props: Props) => {
+  const { section: sectionProp } = props
+  const assessment = useAssessment()
+  const cycle = useCycle()
+  const countryIso = useCountryIso()
+  const dispatch = useAppDispatch()
   const { i18n } = useTranslation()
   const { assessmentName } = useParams<{ assessmentName: AssessmentName; cycleName: string; section: string }>()
-  const assessmentSection = useAssessmentSection()
+  const assessmentSection = useAssessmentSection(sectionProp)
   const tableSections = useTableSections({ sectionName: assessmentSection?.props?.name })
-  const canEditSection = useCanEditSection()
+  const canEditSection = useCanEditSection(sectionProp)
   const { print, onlyTables } = useIsPrint()
 
   const panEuropean = assessmentName === AssessmentName.panEuropean
   const disabled = panEuropean || !canEditSection
+
+  const shouldFetchTableSections = useRef(tableSections.length === 0)
+
+  useEffect(() => {
+    if (shouldFetchTableSections.current) {
+      shouldFetchTableSections.current = false
+      dispatch(
+        AssessmentSectionActions.getTableSections({
+          assessmentName: assessment.props.name,
+          cycleName: cycle.name,
+          section: assessmentSection.props.name,
+          countryIso,
+        })
+      )
+    }
+  }, [tableSections, assessment.props.name, countryIso, cycle.name, dispatch, assessmentSection.props.name, print])
 
   if (!assessmentSection) return null
 
@@ -75,6 +102,10 @@ const AssessmentSection: React.FC = () => {
       <div className="page-break" />
     </div>
   )
+}
+
+AssessmentSection.defaultProps = {
+  section: undefined,
 }
 
 export default AssessmentSection
